Added an exception handler for KeyboardInterrupt.
It was annoying that the program spit out a traceback (With three functions mentioned in it instead of only one!) when interrupting it with [Ctrl]+[C]. So I put the bottom part (which is for importing needed modules and starting the actual program) into an exception handler that just looks for a KeyboardInterrupt error and prevents the program from spitting out a traceback. Every oher error will still be printed.