Update PyPerfScore.py
Added a version note at the beginning. I put the version in a variable and inserted that into the string instead of just putting the version name into the string to make it easier to change the displayed version when it will be shown in several locations in the program. I don't want to miss one version note in the code when stepping up. mismatching version numbers inside the same file would cause a lot of confusion.