Now I started developing the GUI but...
I made a little GUI but that's somehow not workig. I will feed this into my debugger this evening.   
If you want to try this versionn please download it and run it with the commad-line option `--full-gui`, `--debug` can optionally be added too to drop into an interactive shell as soon as the program fails.